Factor out Options and Output
Factored out in order to allow shared use of these classes
